fixed the thing to show which issue you're reading
Before I wasn't sure how to do this but if I just use .btn-info as part
of the jquery selector I don't even have to think about it, nice. So now
some of the buttons are active and some aren't (not all the comics are
done yet) and it doesn't look like you should be able to click on the
ones that you can't.