rendering profiles
so now the functionality is separated between the first and second boxes
and the changes are saved to an appropriate variable, which is then
rendered into the box where the form would be, going to add an undo
button that brings the form back up and then it's time to start tying it
into the combat system again. i think i need to eval the value of the
text boxes in the advanced form, too, so they're integers and not
strings.